+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# Reciept Reader realeses
+
+## 0.0.7
+
+-Adding more fields to the order object. These are: Invoice number and method of payment.
+
+## 0.0.6
+
+-Allowing the widget to be used on flutterflow as flutter flow required a lower sdk version at the time of build
+
 ## 0.0.5
 
 - **Documentation**: Added Dartdoc-style documentation, improving the readability and usability of the widget.